- move files from draft to view/edit
