+++ src/scenes/PlantingSitesRouter/index.tsx
@@ -17,9 +17,9 @@ import PlantingSitesList from './view/PlantingSitesList';
 import PlantingSiteView from './view/PlantingSiteView';
 import PlantingSiteSubzoneView from './view/PlantingSiteSubzoneView';
 import PlantingSiteZoneView from './view/PlantingSiteZoneView';
-import PlantingSiteDraftCreate from './draft/PlantingSiteDraftCreate';
-import PlantingSiteDraftEdit from './draft/PlantingSiteDraftEdit';
-import PlantingSiteDraftView from './draft/PlantingSiteDraftView';
+import PlantingSiteDraftCreate from './edit/PlantingSiteDraftCreate';
+import PlantingSiteDraftEdit from './edit/PlantingSiteDraftEdit';
+import PlantingSiteDraftView from './view/PlantingSiteDraftView';
 import isEnabled from 'src/features';
 
 /**
